Reorganise components for exporting
