+++ packages/webui/src/components/KeyValueList.tsx
@@ -1,4 +1,4 @@
-import { Field } from '@/components/ui';
+import { Field } from './Fields';
 
 export type KeyValuePair = [string, any];
