+++ packages/webui/src/components/SearchInput.tsx
