+++ packages/webui/src/components/index.ts
@@ -1,5 +1,7 @@
 /* istanbul ignore file */
 
+export { default as Authorization } from './Authorization';
+export { default as Button } from './Button';
 export { default as Code } from './Code';
 export { default as DateTime } from './DateTime';
 export { default as DropDown } from './DropDown';
@@ -7,6 +9,7 @@ export { default as Fields, Field } from './Fields';
 export { default as IconButton } from './IconButton';
 export { default as Input } from './Input';
 export { default as JsonDisplay } from './JsonDisplay';
+export { default as KeyValueList } from './KeyValueList';
 export { default as Label } from './Label';
 export { default as Loading } from './Loading';
 export { default as SearchInput } from './SearchInput';
