+++ packages/webui/src/components/ui/RequestHeaders.tsx
@@ -1,8 +1,7 @@
+import Authorization from '@/components/Authorization';
+import KeyValueList from '@/components/KeyValueList';
 import { Trace } from '@/types';
 import { cloneHeaders } from '@/utils';
-
-import Authorization from './Authorization';
-import KeyValueList from './KeyValueList';
 
 export default function RequestHeaders({ trace }: { trace: Trace }) {
   const requestHeaders = trace.http?.requestHeaders;
