+++ packages/webui/src/systems/Sanity.test.tsx
@@ -6,7 +6,7 @@ import { Trace } from '@/types';
 import SanitySystem from './Sanity';
 
 jest.mock(
-  '@/components/ui/Code',
+  '@/components/Code',
   () =>
     function MockCode({ children, ...props }: any) {
       return <div {...props}>{children}</div>;
